# Add templated memcpy/memcmp