--- a/src/common/sort/merge_sorter.cpp
+++ b/src/common/sort/merge_sorter.cpp
@@ -547,8 +547,8 @@ void MergeSorter::MergeData(SortedData &result_data, SortedData &l_data, SortedD
 					const bool &l_smaller = left_smaller[copied + i];
 					const bool r_smaller = !l_smaller;
 					const auto &entry_size = next_entry_sizes[copied + i];
-					memcpy(result_heap_ptr, l_heap_ptr, l_smaller * entry_size);
-					memcpy(result_heap_ptr, r_heap_ptr, r_smaller * entry_size);
+					FastMemcpy(result_heap_ptr, l_heap_ptr, l_smaller * entry_size);
+					FastMemcpy(result_heap_ptr, r_heap_ptr, r_smaller * entry_size);
 					D_ASSERT(Load<uint32_t>(result_heap_ptr) == entry_size);
 					result_heap_ptr += entry_size;
 					l_heap_ptr += l_smaller * entry_size;
@@ -584,8 +584,8 @@ void MergeSorter::MergeRows(data_ptr_t &l_ptr, idx_t &l_entry_idx, const idx_t &
 		const bool &l_smaller = left_smaller[copied + i];
 		const bool r_smaller = !l_smaller;
 		// Use comparison bool (0 or 1) to copy an entry from either side
-		memcpy(target_ptr, l_ptr, l_smaller * entry_size);
-		memcpy(target_ptr, r_ptr, r_smaller * entry_size);
+		FastMemcpy(target_ptr, l_ptr, l_smaller * entry_size);
+		FastMemcpy(target_ptr, r_ptr, r_smaller * entry_size);
 		target_ptr += entry_size;
 		// Use the comparison bool to increment entries and pointers
 		l_entry_idx += l_smaller;