This PR adds templated versions of `memcpy` and `memcmp`, wrapped in methods called `FastMemcpy` and `FastMemcmp`, which select the specific templated version with a large switch statement.

The allows the compiler to pre-compile these functions with the specified number of bytes, and emit more optimized machine code. This yields a speedup specifically when these functions are used in a loop with a constant size, e.g.:
```C++
const size_t size = 8;
for (idx_t i = 0; i < n; i++) {
    // ...
    FastMemcpy(dest, source, size);
    // or
    FastMemcmp(str1, str2, size);
    // ...
}
```

The speedup can be up to 5-6x, depending on `size`.

Of course, these functions are heavily used in the sort code, and adding these functions yields a significant performance boost.
I sorted 100M random integers:
```SQL
.timer on
.mode trash
create table test as select cast(random() * 99999999 as int) i from range(100000000);
-- pragma threads=1; OR pragma threads=4;
select * from test order by i;
```

And here are the results:
| Threads | Old | New |
| :- | :- | :- |
| 1 | 4.285s | **3.966s** |
| 4 | 2.968s | **2.277s** |

As we can see, the performance is mostly felt when there are more threads, because it affects merge sort more than it affects radix sort.

The fun part is that the single threaded performance is now as fast as the 4 threaded performance on this specific benchmark was when [the sorting blog](https://duckdb.org/2021/08/27/external-sorting.html) came out.