@Alex-Monahan 2021-11-23
Wow, incredible that you've made that kind of improvement even after the initial algorithm change! Maybe it's worth re-posting the blog with a short update/new benchmark at the top?