@lnkuiper 2021-11-23
@Alex-Monahan That is a good idea. Maybe I should hold off on that until @Mytherin finishes the local storage rework that will allow for parallel creation of ordered tables?

These two things together will give a great performance boost, especially for strings.